add errors router

+++ modules/error/errorController.js
@@ -1,4 +1,7 @@
+const getEnv = require('../../utils/getEnv');
 const sendRes = require('../../utils/sendRes');
+const asyncHandler = require('express-async-handler');
+const AppError = require('../../utils/AppError');
 
 const deleteErrors = (obj) => {
     if (obj.errors.length == 0) {
@@ -6,18 +9,25 @@ const deleteErrors = (obj) => {
     }
     return obj;
 };
-module.exports = (err, req, res, next) => {
+exports.globalErrorHandler = (err, req, res, next) => {
     const statusCode = err.statusCode || 500;
     const message = err.message || 'Internal Server Error';
-
-    sendRes(
-        res,
-        statusCode,
-        deleteErrors({
-            status: err.status || false,
-            message: message,
-            errors: err.errors || [],
-        }),
-    );
-    next();
+    const NODE_ENV = getEnv('NODE_ENV');
+    if (NODE_ENV == 'dev') {
+        sendRes(
+            res,
+            statusCode,
+            deleteErrors({
+                status: err.status || false,
+                message: message,
+                errors: err.errors || [],
+                stack: err.stack,
+            }),
+        );
+        next();
+    }
 };
+
+exports.customError = asyncHandler(async (req, res, next) => {
+    throw new AppError('error', 300, ['error1', 'error2']);
+});
